requirements.txt
streamlit==1.46.0
face_recognition==1.3.0
face-recognition-models==0.3.0
numpy
Pillow
opencv-python
